refactor(Fluff): Fluff -> Decorations
