+++ Aliucord/src/main/java/com/aliucord/PluginManager.kt
@@ -273,9 +273,9 @@ object PluginManager {
             ButtonsAPI(),
             CommandHandler(),
             CoreCommands(),
+            Decorations(),
             DefaultStickers(),
             ExperimentDefaults(),
-            Fluff(),
             ForwardedMessages(),
             GifPreviewFix(),
             MembersListFix(),
